python/pyproject.toml
[project]
name = "sandbox"
version = "0.1.0"
description = ""
authors = [{ name = "Uy Ha", email = "hchanuy@gmail.com" }]
dependencies = [
    "Jinja2>=3.1.2",
    "tree-sitter>=0.20.2",
    "setuptools>=68.2.2",
    "msgpack>=1.0.7",
    "dacite>=1.8.1",
    "pyserde[msgpack,toml]>=0.13.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
